prueba4
cambio de tipo de dato b-> double

+++ FlashCardsProject/src/ClasePrueba1.java
@@ -1,6 +1,6 @@
 public class ClasePrueba1 {
     //atributos de prueba
     private int a;
-    private int b;
+    private double b;
     private int c;
 }
